first version of command gui for rviz
