+++ cob_command_gui_rviz/lib/plugin.yaml
@@ -1,11 +1,11 @@
-name: cob_rviz_movement_control
-library: cob_rviz_movement_control
+name: cob_command_gui_rviz
+library: cob_command_gui_rviz
 displays:
   -
     class_name: RvizMovementButtons
-    display_name: RvizMovementButtons
+    display_name: cob_command_gui_rviz
     description: |
-      Buttons to control the robot in rviz
+      the cob_command_gui as a rviz plugin
